+++ app_pos/ri_signal/filter/base/filter_base.h
@@ -13,23 +13,23 @@ class SingleSignal;
 
 class FilterBase {
 	public:
-		double Feed(double value);
+		int Feed(int value);
 		virtual void Reset() = 0;
         void AppendFollower(FilterBase* filter);
         void AppendFollower(const char* interest_result_item, FilterBase* filter);
 		bool IsMasterUtility(){return is_master_utility_;}
 		void SetAsMasterUtility(){is_master_utility_ = true;}
-		double GetResult(){return result_value_;}
+		int GetResult(){return result_value_;}
 		
 		virtual void SetKnownPoint(double x, double y) = 0;
 
 
 	protected:
 
-		virtual double ProcessSignal_(double value) = 0;
+		virtual int ProcessSignal_(int value) = 0;
 		std::vector<FilterBase*> followers__;
 		bool is_master_utility_ = false;
-		double result_value_;
+		int result_value_;
 	private:
 };
