+++ app_pos/ri_signal/filter/calculus/differential_time.h
@@ -14,13 +14,13 @@ class DifferentTime : public FilterBase {
 		void SetKnownPoint(double x, double y) override{};
 
 	protected:
-		double ProcessSignal_(double value) override;
+		int ProcessSignal_(int value) override;
 
     private:
 		double sensitivity__ = 1;
 		int feed_count__ = 0;
 		// uint32_t previous_timestamp__;
-		double previous_position__;
+		int previous_position__;
 		// double speed__;
 
 };
